Make FlagDebug public and move it to the model package (#39)
* Make FlagDebug public and move it to the model package

* Add isDebug and respective tests

* Add SampledFlag, refactor flag checking

* Make tests more complicated

* Introduce Flags struct, refactor code to use it on span

* Extend test to use non-zero initialization

+++ model/converter/json/from_domain.go
@@ -59,7 +59,7 @@ func (fd fromDomain) convertSpan(span *model.Span, processID json.ProcessID) jso
 	return json.Span{
 		TraceID:       json.TraceID(span.TraceID.String()),
 		SpanID:        json.SpanID(span.SpanID.String()),
-		Flags:         span.Flags,
+		Flags:         uint32(span.Flags),
 		OperationName: span.OperationName,
 		References:    fd.convertReferences(span),
 		StartTime:     span.StartTime,
